ml/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
catboost==1.2.8
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.117.1
fonttools==4.60.0
graphviz==0.21
h11==0.16.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
narwhals==2.5.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
pyaml==25.7.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0